image-gen-simple/requirements.txt
huggingface_hub
gradio
python-dotenv
pillow
requests
